Added new it in registerAccount.spec.js
